Fix rollover from a ShiftToLayer() qukey with release delay
With qukey defined with an alternate `Key` value of `ShiftToLayer(N)`, we need a way to
send the toggle-off event in order to get the layer to deactivate. Since the physical key
was already released, we need to keep track of the delayed-release qukey, and send the
appropriate release event after the rest of the queue has been flushed.

This change records the delayed key at the time of the delay, and sends the release event
after the ensuing key has been flushed from the queue.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>